Add code page and Unicode range values in OS2 table
This commits adds code for a legacy feature that still is used
by Microsoft Works and probably many other Windows applications
like Microsoft Word.

Birdfont will set the corresponding bit for a Unicode
range if there is at least one character in that range in the font.
Bits for code pages will be set if there is at least one character
other than the default characters.